Refactor & x86_64 Support Attempt

+++ include/Ptrace.h
@@ -9,11 +9,11 @@ bool PtraceStopCallbackResume(int procId, std::function<void()> callback);
 bool SetContext(int procId, user_regs_struct& ctx);
 bool GetContext(int procId, user_regs_struct& ctx);
 
-bool PtraceReadProcessMemory(int pid, unsigned int addr, void* data, size_t len);
-bool PtraceWriteProcessMemory(int pid, unsigned int addr, const void* data, size_t len);
+bool PtraceReadProcessMemory(int pid, uintptr_t addr, void* data, size_t len);
+bool PtraceWriteProcessMemory(int pid, uintptr_t addr, const void* data, size_t len);
 
-bool PtracePushSnapshot(int procId, unsigned int atAddr, size_t len);
-bool PtracePopSnapshot(int procId, unsigned int atAddr);
+bool PtracePushSnapshot(int procId, uintptr_t atAddr, size_t len);
+bool PtracePopSnapshot(int procId, uintptr_t atAddr);
 
 bool PushContext(int procId);
 bool PopContext(int procId);
@@ -21,12 +21,12 @@ bool PopContext(int procId);
 bool PtraceContinue(int procId);
 
 /*This function spect a context of ptrace alredy attached and the process alredy stopped*/
-uintptr_t PtraceCall(int procId, uintptr_t entry, const std::vector<uint32_t>& params);
-uintptr_t PtraceCallModuleSymbol(int procId, const char* module, const char* symbol, bool nb, const std::vector<uint32_t>& params);
+uintptr_t PtraceCall(int procId, uintptr_t entry, const std::vector<size_t>& params);
+uintptr_t PtraceCallModuleSymbol(int procId, const char* module, const char* symbol, bool nb, const std::vector<size_t>& params);
 
 /* Wrappers */
 template<typename T>
-T PtraceReadProcessMemoryWrapper(int procId, unsigned int addr)
+T PtraceReadProcessMemoryWrapper(int procId, uintptr_t addr)
 {
     T obj;
 
@@ -36,7 +36,7 @@ T PtraceReadProcessMemoryWrapper(int procId, unsigned int addr)
 }
 
 template<typename T>
-bool PtraceWriteProcessMemoryWrapper(int procId, unsigned int addr, const T& obj)
+bool PtraceWriteProcessMemoryWrapper(int procId, uintptr_t addr, const T& obj)
 {
     return PtraceWriteProcessMemory(procId, addr, (const void*)&obj, sizeof(T));
 }
